--- game/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Assessment Race</title>
-    <link rel="stylesheet" href="./css/style.css" />
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.11.5/gsap.min.js"></script>
-    <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/110/three.min.js"></script> -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/simplex-noise/2.4.0/simplex-noise.min.js"></script>
-    <script src="./js/pixi.js"></script>
-</head>
-
-<body>
-    <div class="score-container">
-        <div id="score"></div>
-    </div>
-
-    <div id="scene"></div>
-
-    <script src="./js/scripts.js" type="module"></script>
-</body>
-
-</html>
