--- index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./style.css" />
-    <title>Exploring Code</title>
-  </head>
-  <body>
-    <div class="container">
-      <h1>Welcome to Our Website</h1>
-      <div class="demo">
-        <div>CANVAS</div>
-        <div>THREE.JS</div>
-        <div>WEB AUDIO API</div>
-        <div>Gsap Animation</div>
-      </div>
-      <div class="button-groups">
-        <a href="#">Enter</a>
-      </div>
-    </div>
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.11.5/gsap.min.js"></script>
-    <script src="./script.js"></script>
-  </body>
-</html>
