--- lose/style.css
@@ -1,24 +1,0 @@
-*{
-	padding: 0;
-	margin: 0;
-	box-sizing: border-box;
-}
-
-canvas{
-background-color: black;
-position: absolute;
-top: 0;
-left: 0;
-}
-
-button {
-  position: absolute;
-  top: 20px;
-  left: 20px;
-  padding: 10px 20px;
-  background-color: #f2f2f2;
-  color: #333;
-  font-size: 16px;
-  border: none;
-  cursor: pointer;
-}
