+++ package.json
@@ -18,14 +18,15 @@
     "url": "https://github.com/desmati/assessment-race/issues"
   },
   "homepage": "https://github.com/desmati/assessment-race#readme",
-	"devDependencies": {
-		"@parcel/config-default": "^2.4.0",
+  "devDependencies": {
+    "@parcel/config-default": "^2.4.0",
     "parcel-reporter-static-files-copy": "^1.3.4"
-	},
-	"dependencies": {
-		"three": "^0.139.0",
+  },
+  "dependencies": {
+    "parcel": "^2.9.2",
+    "three": "^0.139.0",
     "yuka": "^0.7.7"
-	},
+  },
   "staticFiles": {
     "staticOutPath": "assets"
   }
